Fix MAR, finalize the inference, and revise the output
MAR formula is fixed to follow dlib face point. Inference is finalized to give output for all drowsiness indications that show 60% out of total frames for the last 10 seconds. If the driver is drowsy, the output will show drowsiness, reason, and suggest rest. If the driver is OK, the output will show OK.